feat(witnessgen): collect impl.rs to types
