+++ prover/src/utils.rs
@@ -1,7 +1,7 @@
 use alloy_primitives::B256;
 use anyhow::Result;
 use kroma_utils::deps_version::SP1_SDK_VERSION;
-use kroma_witnessgen::get_witness_impl::WitnessResult;
+use kroma_witnessgen::types::WitnessResult;
 use sp1_sdk::network::client::NetworkClient;
 use sp1_sdk::proto::network::{ProofMode, ProofStatus};
 use sp1_sdk::{block_on, SP1ProofWithPublicValues, SP1Stdin};
