+++ witnessgen/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod errors;
-pub mod get_witness_impl;
 pub mod interface;
-pub mod request_witness_impl;
-pub mod spec_impl;
+pub mod types;
+pub mod utils;
 pub mod witness_db;
